Completed loading/saving tasks, added timestamps
