--- tests/default_example_diary_17082023.json
@@ -1,17 +1,0 @@
-{
-  "date": "17.08.2023",
-  "activity": [],
-  "important": [],
-  "focus": 3,
-  "starting_mood": 3,
-  "ending_mood": 3,
-  "satisfaction": 3,
-  "tags": [],
-  "focus_comment": null,
-  "start_mood_comment": null,
-  "end_mood_comment": null,
-  "satisfaction_comment": null,
-  "comment": null,
-  "important_comment": false,
-  "changelog": []
-}
